requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
sentence-transformers
faiss-cpu
scikit-learn